+++ lib/src/canvas_terminal.dart
@@ -14,28 +14,37 @@ class CanvasTerminal extends RenderableTerminal {
 
   final Font _font;
   final html.CanvasElement _canvas;
-  html.CanvasRenderingContext2D _context;
+  final html.CanvasRenderingContext2D _context;
 
-  int _scale = 1;
+  /// The drawing scale, used to adapt to Retina displays.
+  final int _scale = html.window.devicePixelRatio.toInt();
 
   Vec get size => _display.size;
   int get width => _display.width;
   int get height => _display.height;
 
-  CanvasTerminal(int width, int height, this._canvas, this._font)
-      : _display = new Display(width, height) {
-    _context = _canvas.context2D;
+  factory CanvasTerminal(int width, int height, Font font, [html.CanvasElement canvas]) {
+    var display = new Display(width, height);
 
-    _canvas.width = _font.charWidth * width;
-    _canvas.height = _font.lineHeight * height;
-
-    // Handle high-resolution (i.e. retina) displays.
-    if (html.window.devicePixelRatio > 1) {
-      _scale = 2;
-
-      _canvas.style.width = '${_font.charWidth * width / _scale}px';
-      _canvas.style.height = '${_font.lineHeight * height / _scale}px';
+    // If not given a canvas, create one and add it to the page.
+    if (canvas == null) {
+      canvas = new html.CanvasElement();
+      html.document.body.append(canvas);
     }
+
+    return new CanvasTerminal._(display, font, canvas);
+  }
+
+  CanvasTerminal._(this._display, this._font, html.CanvasElement canvas)
+      : _canvas = canvas,
+        _context = canvas.context2D {
+    // Handle high-resolution (i.e. retina) displays.
+    var canvasWidth = _font.charWidth * _display.width;
+    var canvasHeight = _font.lineHeight * _display.height;
+    _canvas.width = canvasWidth * _scale;
+    _canvas.height = canvasHeight * _scale;
+    _canvas.style.width = '${canvasWidth}px';
+    _canvas.style.height = '${canvasHeight}px';
   }
 
   void drawGlyph(int x, int y, Glyph glyph) {
@@ -50,15 +59,17 @@ class CanvasTerminal extends RenderableTerminal {
 
       // Fill the background.
       _context.fillStyle = glyph.back.cssColor;
-      _context.fillRect(x * _font.charWidth, y * _font.lineHeight,
-          _font.charWidth, _font.lineHeight);
+      _context.fillRect(x * _font.charWidth * _scale,
+          y * _font.lineHeight * _scale,
+          _font.charWidth * _scale, _font.lineHeight * _scale);
 
       // Don't bother drawing empty characters.
       if (char == 0 || char == CharCode.SPACE) return;
 
       _context.fillStyle = glyph.fore.cssColor;
       _context.fillText(new String.fromCharCodes([char]),
-          x * _font.charWidth + _font.x, y * _font.lineHeight + _font.y);
+          (x * _font.charWidth + _font.x) * _scale,
+          (y * _font.lineHeight + _font.y) * _scale);
     });
   }